+++ CHANGELOG.md
@@ -1,15 +1,46 @@
-## [1.0.0] - 30.06.2018
+## [2.0.7] - 18.09.2018
 
-* Initial release.
+* Added:
+```Dart
+/// [serverSideConstraints] : If you need some serverside filtering besides the [Area] pass a list of [QueryConstraint]
+/// [serverSideOrdering] : If you need some serverside ordering you can pass a List of [OrderConstraints]
+```
+To `getDataInArea()`
 
-## [2.0.0] - 03.07.2018
+## [2.0.6] - 06.09.2018
 
-* Breaking changes in naming of Area members and adding more functions
+* Updated dependency to cloud_firestore 0.8.0## 
+
+## [2.0.5] - 07.08.2018
+
+* `buildQuery` now accepts a Query as data source, before only DocumentCollections were possible.
+
+## [2.0.4] - 04.08.2018
+
+* `getDataInArea` now accepts a Query as data source, before only DocumentCollections were possible.
+
+## [2.0.3] - 11.07.2018
+
+* Fixed a bug in distance calculation.
 
 ## [2.0.2] - 10.07.2018
 
 * If you provide a `datalocationAccessor` to `getDataInArea` the results are not additionally filtered so that you only get places within the given radius not in a square. Also some bug fixes.
 
-## [2.0.3] - 11.07.2018
+## [2.0.0] - 03.07.2018
 
-* Fixed a bug in distance calculation.
+* Breaking changes in naming of Area members and adding more functions
+
+##[1.0.0] - 30.06.2018
+
+* Initial release.
+
+
+
+
+
+
+
+
+
+